# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Event Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /events/{id}/register:
    post:
      operationId: registerForEvent
      tags: [registrations]
      summary: Register for an event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
  /registrations/my:
    get:
      operationId: getMyRegistrations
      tags: [registrations]
      summary: Get my registered events
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: My registrations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationListResponse"
  /registrations/{id}:
    delete:
      operationId: cancelRegistration
      tags: [registrations]
      summary: Cancel a registration
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cancelled
  /registrations/{id}/qr:
    get:
      operationId: getRegistrationQr
      tags: [registrations]
      summary: Get QR code for ticket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCodeResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, organizer, participant]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        isActive:
          type: boolean
    UserListResponse:
      type: object
      required: [users, total, page, totalPages]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    Event:
      type: object
      required: [id, title, description, date, venue, category, capacity, registrationDeadline, status, approvalStatus, organizerId, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
        venue:
          type: string
        category:
          type: string
        capacity:
          type: integer
        registrationDeadline:
          type: string
        status:
          type: string
          enum: [upcoming, ongoing, completed]
        approvalStatus:
          type: string
          enum: [pending, approved, rejected]
        bannerUrl:
          type: ["string", "null"]
        organizerId:
          type: integer
        organizerName:
          type: ["string", "null"]
        registrationCount:
          type: integer
        isRegistered:
          type: boolean
        createdAt:
          type: string
...
    Registration:
      type: object
      required: [id, eventId, userId, registeredAt, attended]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        userId:
          type: integer
        registeredAt:
          type: string
        attended:
          type: boolean
        qrCode:
          type: ["string", "null"]
    RegistrationListResponse:
      type: object
      required: [registrations, total, page, totalPages]
      properties:
        registrations:
          type: array
          items:
            $ref: "#/components/schemas/RegistrationWithEvent"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    RegistrationWithEvent:
      type: object
      required: [id, eventId, userId, registeredAt, attended]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        userId:
          type: integer
        registeredAt:
          type: string
        attended:
          type: boolean
        qrCode:
          type: ["string", "null"]
        event:
          $ref: "#/components/schemas/Event"
    QrCodeResponse:
      type: object
      required: [qrCode, registrationId, eventTitle]
      properties:
        qrCode:
          type: string
        registrationId:
          type: integer
        eventTitle:
          type: string
...