# lib/api-zod/src/generated/types/registrationListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Event Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationWithEvent } from './registrationWithEvent';

export interface RegistrationListResponse {
  registrations: RegistrationWithEvent[];
  total: number;
  page: number;
  totalPages: number;
}